Viewmodels now expose UI error classes instead of domain errors
