+++ feature/src/main/java/com/example/feature/countrydetails/CountryDetailsErrors.kt
@@ -3,23 +3,9 @@ package com.example.feature.countrydetails
 import android.content.Context
 import com.example.errors.CountryDetailsError
 import com.example.features.R
-import com.example.uicomponents.error.UiErrorMapper
-
-sealed class CountryDetailsUiError
-sealed class CountryDetailsFullScreenError : CountryDetailsUiError() {
-    object CountryNotFound : CountryDetailsFullScreenError()
-    data class Generic(val message: String?) : CountryDetailsFullScreenError()
-}
-
-object CountryDetailsUiErrorMapper : UiErrorMapper<CountryDetailsError, CountryDetailsUiError> {
-
-    override fun toUiError(error: CountryDetailsError): CountryDetailsUiError {
-        return when (error) {
-            CountryDetailsError.CountryNotFound -> CountryDetailsFullScreenError.CountryNotFound
-            is CountryDetailsError.Other -> CountryDetailsFullScreenError.Generic(error.message)
-        }
-    }
-}
+import com.example.viewmodels.error.countrydetails.CountryDetailsFullScreenError
+import com.example.viewmodels.error.countrydetails.CountryDetailsUiError
+import com.example.viewmodels.error.countrydetails.CountryDetailsUiErrorMapper
 
 class CountryDetailsFullScreenErrorMessageFactory(private val context: Context) {
 
@@ -31,6 +17,5 @@ class CountryDetailsFullScreenErrorMessageFactory(private val context: Context) 
     }
 }
 
-fun CountryDetailsError.toUiError(): CountryDetailsUiError = CountryDetailsUiErrorMapper.toUiError(this)
 fun CountryDetailsFullScreenError.message(context: Context): String =
     CountryDetailsFullScreenErrorMessageFactory(context).getMessage(this)
