+++ feature/src/main/java/com/example/feature/countrydetails/CountryDetailsPage.kt
@@ -9,12 +9,13 @@ import com.example.feature.countrydetails.component.CountryDetailsContent
 import com.example.feature.countrydetails.component.FullScreenErrorView
 import com.example.uicomponents.library.ProgressIndicator
 import com.example.viewmodels.CountryDetailsViewModel
+import com.example.viewmodels.error.countrydetails.CountryDetailsFullScreenError
 
 @Composable
 fun CountryDetailsPage(detailsUIState: CountryDetailsViewModel.UiState) {
     ProgressIndicator(isLoading = detailsUIState is CountryDetailsViewModel.UiState.Loading)
     if (detailsUIState is CountryDetailsViewModel.UiState.Error) {
-        when (val uiError = detailsUIState.error.toUiError()) {
+        when (val uiError = detailsUIState.error) {
             is CountryDetailsFullScreenError -> FullScreenErrorView(uiError.message(LocalContext.current))
         }
     }
