+++ viewmodels/src/main/java/com/example/viewmodels/error/UiErrorMapper.kt
@@ -1,0 +1,5 @@
+package com.example.viewmodels.error
+
+interface UiErrorMapper<LogicError, UiError> {
+    fun toUiError(error: LogicError): UiError?
+}
