+++ viewmodels/src/main/java/com/example/viewmodels/error/countrydetails/CountryDetailsUiError.kt
@@ -1,0 +1,7 @@
+package com.example.viewmodels.error.countrydetails
+
+sealed class CountryDetailsUiError
+sealed class CountryDetailsFullScreenError : CountryDetailsUiError() {
+    object CountryNotFound : CountryDetailsFullScreenError()
+    data class Generic(val message: String?) : CountryDetailsFullScreenError()
+}
