+++ viewmodels/src/main/java/com/example/viewmodels/error/countrylist/CountryListUiError.kt
@@ -1,4 +1,4 @@
-package com.example.feature.countrylist.error
+package com.example.viewmodels.error.countrylist
 
 sealed class CountryListUiError
